# Conflicts:
#	portlets/message/src/main/java/com/aimluck/eip/message/MessageRoomFormData.java
#	sql/postgres/org001.sql

+++ portlets/webmail/src/main/java/com/aimluck/eip/webmail/WebMailAccountFormData.java
@@ -34,6 +34,8 @@ import com.aimluck.commons.field.ALStringField;
 import com.aimluck.commons.utils.ALStringUtil;
 import com.aimluck.eip.cayenne.om.portlet.EipMMailAccount;
 import com.aimluck.eip.cayenne.om.portlet.EipTMail;
+import com.aimluck.eip.cayenne.om.portlet.EipTMailFilter;
+import com.aimluck.eip.cayenne.om.portlet.EipTMailFolder;
 import com.aimluck.eip.common.ALAbstractFormData;
 import com.aimluck.eip.common.ALDBErrorException;
 import com.aimluck.eip.common.ALEipConstants;
@@ -47,7 +49,6 @@ import com.aimluck.eip.modules.actions.common.ALAction;
 import com.aimluck.eip.modules.screens.WebMailAdminFormJSONScreen;
 import com.aimluck.eip.modules.screens.WebMailAdminFormScreen;
 import com.aimluck.eip.orm.Database;
-import com.aimluck.eip.orm.query.SQLTemplate;
 import com.aimluck.eip.orm.query.SelectQuery;
 import com.aimluck.eip.services.eventlog.ALEventlogConstants;
 import com.aimluck.eip.services.eventlog.ALEventlogFactoryService;
@@ -451,7 +452,7 @@ public class WebMailAccountFormData extends ALAbstractFormData {
   }
 
   /**
-   * 
+   *
    * @param rundata
    * @param context
    * @param msgList
@@ -516,7 +517,7 @@ public class WebMailAccountFormData extends ALAbstractFormData {
   }
 
   /**
-   * 
+   *
    * @param rundata
    * @param context
    * @param msgList
@@ -556,7 +557,7 @@ public class WebMailAccountFormData extends ALAbstractFormData {
   }
 
   /**
-   * 
+   *
    * @param rundata
    * @param context
    * @param msgList
@@ -644,7 +645,7 @@ public class WebMailAccountFormData extends ALAbstractFormData {
   }
 
   /**
-   * 
+   *
    * @param rundata
    * @param context
    * @param msgList
@@ -683,12 +684,24 @@ public class WebMailAccountFormData extends ALAbstractFormData {
       Database.delete(account);
       Database.commit();
 
-      // delete from database
+      // delete from database eip_t_mail, eip_t_mail_folder, eip_t_mail_filter
+      String filterSql =
+        "DELETE FROM eip_t_mail_filter WHERE account_id = #bind($accountId)";
+      Database.sql(EipTMailFilter.class, filterSql).param(
+        "accountId",
+        account.getAccountId()).execute();
+
+      String folderSql =
+        "DELETE FROM eip_t_mail_folder WHERE account_id = #bind($accountId)";
+      Database.sql(EipTMailFolder.class, folderSql).param(
+        "accountId",
+        account.getAccountId()).execute();
+
       String sql =
-        "DELETE FROM eip_t_mail WHERE account_id = "
-          + String.valueOf(account.getAccountId());
-      SQLTemplate<EipTMail> sqlTemplate = Database.sql(EipTMail.class, sql);
-      sqlTemplate.execute();
+        "DELETE FROM eip_t_mail WHERE account_id = #bind($accountId)";
+      Database.sql(EipTMail.class, sql).param(
+        "accountId",
+        account.getAccountId()).execute();
 
       // セッション変数を削除する
       WebMailUtils.clearWebMailAccountSession(rundata, context);
